CLN : literalinclude key bindings example
	new file:   docs/source/traitsui_user_manual/examples/key_bindings.py
	modified:   docs/source/traitsui_user_manual/factories_advanced_extra.rst
	deleted:    examples/tutorials/doc_examples/examples/key_bindings.py
